Prep v2.0.0 Release
This involved a handful of boring changes:
1. Updated release number from 0.1.0 -> 2.0.0
2. Updated name of project `gradle-test` -> `nusports`
3. Moved package structure `github.ptrteixeira` ->
`com.github.ptrteixeira.nusports`

Preparing for tagging and releasing.
