+++ build.gradle
@@ -1,5 +1,5 @@
 buildscript {
-  ext.kotlin_version = '1.0.3'
+  ext.kotlin_version = '1.0.4'
   ext.junit_version = '1.0.0-M2'
 
   repositories {
@@ -26,8 +26,8 @@ apply plugin: 'kotlin'
 apply from: "gradle/style.gradle"
 apply from: "gradle/uitest.gradle"
 
-group 'github.PtrTeixeira'
-version '0.1.0'
+group 'com.com.github.PtrTeixeira'
+version '2.0.0'
 
 def versions = [
    jsoup: '1.9.2',
