+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'gradle-test'
+rootProject.name = 'NUSports'
