+++ src/main/java/com/github/ptrteixeira/nusports/model/Site.java
@@ -1,0 +1,10 @@
+package com.github.ptrteixeira.nusports.model;
+
+/**
+ * Enumeration of websites for which web scrapers have been implemented.
+ *
+ * @author Peter Teixeira
+ */
+public enum Site {
+  CAA
+}
