--- src/main/java/github/ptrteixeira/model/Site.java
@@ -1,9 +1,0 @@
-package github.ptrteixeira.model;
-
-/**
- * Enumeration of websites for which web scrapers have been implemented.
- * @author Peter Teixeira
- */
-public enum Site {
-  CAA
-}
